Add a wide-column db as a new data storage concept (#95)
# High Level Overview

Proposal to add a wide-column db as a new data storage. I attempted to
draw a distinction between a column-oriented db and a wide-column-db
(although it depends on the source). I thought it would be valuable as I
got confused with this myself.

Happy for corrections/clarifications/feedback - even if they are spicy
haha!